Backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anyio==3.7.1
motor==3.7.1
pymongo==4.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pydantic==2.7.4
pydantic-settings==2.10.1
python-multipart==0.0.20
email-validator==2.2.0
google-auth==2.40.3
openai-whisper==20250625
langdetect==1.0.9
ffmpeg-python==0.2.0
torch==2.3.1
transformers==4.39.3
sentencepiece==0.2.0
sacrebleu==2.4.2
numba==0.58.1
llvmlite==0.41.1
indic-nlp-library==0.92
numpy<2.0.0
python-dotenv==1.0.1
websockets==11.0.3
tiktoken==0.7.0
regex==2023.12.25
pytz==2024.1
tzdata==2024.1
sentence-transformers
indic-transliteration
sacremoses
accelerate
huggingface_hub
faiss-cpu
PyPDF2
librosa
soundfile
